Rename the Folder name to ModeButtom from ModeOption
